+++ src/buildTask.ts
@@ -9,41 +9,73 @@ export class BuildTask {
         }
     }
 
+    /** The name of the task. */
     public readonly name: string;
+    /** whether the task has run. */
     public hasRun: boolean = false;
+    /** The result of the task. */
     public result: any;
-    public successful?: boolean;
+    /** Tasks this task depends on. */
     public dependencies: string[] = [];
+    /** Tasks that run as part of this task. */
     public runTasks: string[] = [];
+    /** The error of the task. */
     public error: any;
+    /** When to run the task. */
     public whenReference: BuildTaskWhen | WhenPredicate = BuildTaskWhen.SuccessfulDependencies;
+    /** What the task does. */
     public doesReference?: DoesDelegate;
+    /** What to run when an error occurs. */
     public errorReference?: ErrorDelegate;
 
+    /**
+     * Creates a new task.
+     * @param name The name of the task.
+     */
     constructor(name: string) {
         this.name = name;
     }
 
+    /**
+     * Controls under what circumstances the task will run.
+     * @param when When the task will run.
+     */
     public when(when: BuildTaskWhen | WhenPredicate): BuildTask {
         this.whenReference = when;
         return this;
     }
 
+    /**
+     * A set of tasks that must run before this task can run.
+     * @param tasks Dependent tasks.
+     */
     public dependsOn(...tasks: Array<string | BuildTask>): BuildTask {
         BuildTask.addTasks(this.dependencies, ...tasks);
         return this;
     }
 
+    /**
+     * Sets what the task does.
+     * @param method What the task does.
+     */
     public does(method: DoesDelegate): BuildTask {
         this.doesReference = method;
         return this;
     }
 
+    /**
+     * A set of tasks that will run as part of this task.
+     * @param tasks Tasks to run.
+     */
     public runs(...tasks: Array<string | BuildTask>): BuildTask {
         BuildTask.addTasks(this.runTasks, ...tasks);
         return this;
     }
 
+    /**
+     * Sets what to run when an error occurs.
+     * @param callback Method to handle errors.
+     */
     public onError(callback: ErrorDelegate): BuildTask {
         this.errorReference = callback;
         return this;
@@ -55,6 +87,8 @@ export type ErrorDelegate = (err: any) => any;
 export type WhenPredicate = (task: BuildTask) => boolean;
 
 export enum BuildTaskWhen {
+    /** Run only when all dependent tasks have completed without error. */
     SuccessfulDependencies,
+    /** Runs regardless of dependent tasks. */
     Always,
 }